+++ lib/recording_studio_publications.rb
@@ -2,7 +2,10 @@
 
 require "recording_studio"
 require "recording_studio_publications/version"
-require "recording_studio_publications/logo_authorization"
+require "recording_studio_publications/publication_type"
+require "recording_studio_publications/family_authorization"
+require "recording_studio_publications/family_management"
+require "recording_studio_publications/published_publication"
 require "recording_studio_publications/engine"
 require "recording_studio_publications/configuration"
 require "recording_studio_publications/capabilities/example"